feat: add unescape for user and password
To unescape the user and password in the URI, thereby expanding the character set supported in the user and password fields.

The HTTP standard inherently supports unescaping in URIs to accommodate a broader range of characters.
When ostree calls g_uri_parse, if the G_URI_FLAGS_ENCODED parameter is not included, it automatically performs unescaping.
Theoretically, removing G_URI_FLAGS_ENCODED could achieve the desired outcome.

However, it is uncertain whether removing G_URI_FLAGS_ENCODED might affect other processes. Therefore, while retaining the existing logic,
it is decided to manually invoke the unescaping function specifically for the user and password sections.